06/11/23-17:16 Guito

+++ src/main/java/com/sisu/sisu/Dao/DipDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.Dip;
 
-public interface DipDao extends CrudRepository<Dip, Long>{
+public interface DipDao extends CrudRepository<Dip, Integer>{
     
 }
